% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setParam.R
\name{buildDLM}
\alias{buildDLM}
\title{A function to retun the required covariance matrices to compute the transition probabilities of the HMDP.}
\usage{
buildDLM(iniDLM, param, Y)
}
\arguments{
\item{iniDLM}{A list that contains the initial parameters of the related DLM.}

\item{param}{A list including the parameters of the HMDP model.}

\item{Y}{A list included the observations for the related DLM.}
}
\value{
A list containing covariance matrices used in file dlm.h to compute the transition probabilities.
}
\description{
A function to retun the required covariance matrices to compute the transition probabilities of the HMDP.
}
\author{
Reza Pourmoayed \email{rpourmoayed@econ.au.dk}
}

