% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BuildHMDP2}
\alias{BuildHMDP2}
\title{Build the HMDP (2 levels with shared linking) using the C++ binary writer.}
\usage{
BuildHMDP2(filePrefix, param, paramDLMP, paramDLMPi, paramDLMF, paramPolicy)
}
\arguments{
\item{filePrefix}{Prefix used by the binary files storing the MDP model.}

\item{param}{Model parameters of the HMDP.}

\item{paramDLMP}{A list including the parameters for the DLMP (DLM for pork price information).}

\item{paramDLMPi}{A list including the parameters for the DLMW (DLM for weigh information).}

\item{paramDLMF}{A list including the parameters for the DLMF (DLM for feed price information)}

\item{paramPolicy}{A given policy for the hmdp. In finding the optimal policy of the HMDP, this policy will not be used.}
}
\value{
Build log (character).
}
\description{
Build the HMDP (2 levels with shared linking) using the C++ binary writer.
}
\author{
Lars Relund \email{lars@relund.dk}, Reza Pourmoayed \email{rpourmoayed@econ.au.dk}
}

