% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setParam.R
\name{findIndex}
\alias{findIndex}
\title{A function to find the id of a state in the discretized intervals.}
\usage{
findIndex(st, matDis)
}
\arguments{
\item{st}{Given value of state.}

\item{matDis}{A matrix including the discretized intervals of the related stae.}
}
\value{
State id in matDis.
}
\description{
A function to find the id of a state in the discretized intervals.
}
\author{
Reza Pourmoayed \email{rpourmoayed@econ.au.dk}
}

