% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/packageInfo.R
\docType{package}
\name{hmdpPricePigIT-package}
\alias{PigIT}
\alias{hmdpPricePigIT-package}
\title{Slaughter pig marketing under price fluctuations}
\description{
This package includes models and tools used to find the best marketing policy of a pig farm
under price fluctuations. The model is developed under the PigIT project \url{www.pigit.net}.
}
\section{History}{

\itemize{
  \item
  \item 2015-04: The model was coded using external processes to save the memory usage.
  \item 2015-08: New simulatiom merhos was used to find the rewards of actions based on the given RRM.
  \item 2013-09: A new part was added to the codes to calculate the value of price information
}
}
\author{
Lars Relund \email{lars@relund.dk} and Reza Pourmoayed \email{rpourmoayed@econ.au.dk}
}

