% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setParam.R
\name{paramDLMs}
\alias{paramDLMs}
\title{Initialize the parameters for the DLM used for the weight (DLMW), pig price (DLMP) and feed price (DLMF).
We assume the FF and GG matrices of these DLMs are time varying.}
\usage{
paramDLMs(dlmPig, dlmFeed, dlmPiglet)
}
\arguments{
\item{dlmPig}{DLM related to the carcass price}

\item{dlmFeed}{DLM related to the feed price}

\item{dlmPiglet}{DLM related to the piglet price}
}
\value{
A list containing all the parameters for three DLMs: DLMW, DLMP, DLMF
}
\description{
Initialize the parameters for the DLM used for the weight (DLMW), pig price (DLMP) and feed price (DLMF).
We assume the FF and GG matrices of these DLMs are time varying.
}
\author{
Reza Pourmoayed \email{rpourmoayed@econ.au.dk}
}

