% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pNorm1D}
\alias{pNorm1D}
\title{Univariate cumulative normal (distribution function)}
\usage{
pNorm1D(x, mean, sd)
}
\arguments{
\item{x}{Value of quantiles.}

\item{mean}{The mean.}

\item{sd}{The standard deviation.}
}
\value{
The distribution function.
}
\description{
Based on the paper "Better approximations to cumulative normal functions" by 
G. West \url{http://www.wilmott.com/pdfs/090721_west.pdf}. The source code
has been taken from the authors webpage \url{http://finmod.co.za/research.html}.
}
\author{
Lars Relund \email{lars@relund.dk}
}

