% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pNorm2D}
\alias{pNorm2D}
\title{Bivariate cumulative normal (distribution function)}
\usage{
pNorm2D(lower, upper, mean, sigma)
}
\arguments{
\item{lower}{The vector of lower limits (2-dim vector).}

\item{upper}{The vector of upper limits (2-dim vector).}

\item{mean}{The mean vector (2-dim vector).}

\item{sigma}{The covariance matrix (2x2).}
}
\value{
The distribution function.
}
\description{
Based on the paper "Better approximations to cumulative normal functions" by
G. West \url{http://www.wilmott.com/pdfs/090721_west.pdf}. The source code
has been taken from the authors webpage \url{http://finmod.co.za/research.html}.
}
\details{
Transform into a bivariate with sd = 1 using Y = 1/sqrt(diag(sigma))(X-mean), i.e.
Pr(X<=x) = P(Y<=1/sqrt(diag(sigma))(x-mean)).
}
\author{
Lars Relund \email{lars@relund.dk}
}

