% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SimulatePigs}
\alias{SimulatePigs}
\title{Calculate the reward of seeling and the feed intake of the pigs in an unselected pen using simulation.}
\usage{
SimulatePigs(pigs, samples, weeks, prices, V, B, R)
}
\arguments{
\item{pigs}{Number of pigs in the pen.}

\item{samples}{Number of samples in the simulation.}

\item{weeks}{Number of weeks in a production cycle.}

\item{prices}{Set of possible prices for the carcass.}

\item{V}{Covariance matrix of the RRM model for the random parameters.}

\item{B}{Fixed parameters of the RRM model.}

\item{R}{Standard deviation of residual error in the RRM model.}
}
\value{
A list used in R including the weight, growth, leanness,
and the reward of sorted pigs in the pen.
}
\description{
Calculate the reward of seeling and the feed intake of the pigs in an unselected pen using simulation.
}

