% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setParam.R
\name{setParameters}
\alias{setParameters}
\title{Set the parameters used when build the HMDP model}
\usage{
setParameters(tMax = 15, tStartMarketing = 9, pigs = 15,
  minMarketingSize = 0, avgGRate = 6, avgLeanP = 61, avgInsWeight = 30,
  avgInsSd = 2.2, convRateSd = 1.4, cleaningPeriod = 4,
  marketingLength = 3, pigletLeadTime = 1, modPolicy = FALSE,
  rewCull = list(), feedWeek = matrix(), feedCull = matrix(), iMTP = 2,
  iMTF = 2, iMSP = 2, iMSF = 2, iMSPi = 1,
  meanWeights = seq(avgInsWeight, 144, by = 6),
  sdWeights = round(seq(avgInsSd, 13, length = tMax), 2),
  meanGrowth = seq(5.5, 9, by = 0.2), sdGrowth = seq(1, 2.5, by = 0.1),
  centerPointsTP = round(seq(8, 13, length = 8), 2),
  centerPointsTF = round(seq(1.1, 2.2, length = 6), 2),
  centerPointsSP = round(seq(-0.4, 0.4, length = 5), 2),
  centerPointsSPi = round(seq(3.45, 3.65, length = 4), 2),
  centerPointsSF = round(seq(-0.2, 0.2, length = 5), 2))
}
\arguments{
\item{tMax}{Maximum growth period in weeks.}

\item{tStartMarketing}{Week where start to consider marketing.}

\item{pigs}{Number of pigs in a pen at insertion.}

\item{minMarketingSize}{Desired minimum marketing group.}

\item{avgGRate}{Average growth rate in the herd.}

\item{avgLeanP}{Average leanness percent in the herd.}

\item{avgInsWeight}{Avarage weight of the pigs at the insertion time into the pen.}

\item{avgInsSd}{Avarage of standard deviation of weight of the pigs at the insertion time into the pen.}

\item{convRateSd}{Sd of conversion rate between live weight and slaughter weight.}

\item{cleaningPeriod}{Cleaninig period after terminating the pen.}

\item{marketingLength}{Daily length from marketing action to sending the pigs to slaughterhouse.}

\item{pigletLeadTime}{Lead time length of buying new piglets. Related decision is made in termination time.}

\item{modPolicy}{a bool variable showing the HMDP will be solved for a given policy.}

\item{rewCull}{A list containing the rewards of culling 1 pig according to the different pork prices.}

\item{feedWeek}{A matrix containing the feed intake of sorted pigs for one 1 week.}

\item{feedCull}{A matrix containing the feed intake of sorted pigs for marketingLength days.}

\item{iMTP}{id of pork price in a given modified policy (without price deviation)}

\item{iMTF}{id of feed price in a given modified policy (without price deviation)}

\item{iMSP}{id of pork price deviation in a given modified policy (without price deviation)}

\item{iMSF}{id of feed price deviation in a given modified policy (without price deviation)}

\item{iMSPi}{id of piglet price deviation in a given modified policy (without price deviation)}

\item{meanWeights}{Avergae weight of the pigs in the RRM model.}

\item{sdWeights}{Standard deviation of weights during the growing period in the RRM model.}

\item{meanGrowth}{Average growth of the pigs in the RRM model.}

\item{sdGrowth}{Standard deviation of growth during the growing period in the RRM model.}

\item{centerPointsTP}{Center points used in the discretization for the trend of pig price estimate.}

\item{centerPointsTF}{Center points used in the discretization for the trend of feed price estimate.}

\item{centerPointsSP}{Center points used in the discretization for the slope of pig price estimate.}

\item{centerPointsSPi}{Center points used in the discretization for the slope of piglet price estimate.}

\item{centerPointsSF}{Center points used in the discretization for the slope of feed price estimate.}
}
\value{
A list containing all the parameters
}
\description{
Set the parameters used when build the HMDP model
}
\author{
Reza Pourmoayed \email{rpourmoayed@econ.au.dk} and Lars Relund \email{lars@relund.dk}
}

